+++ liborbital/point.rs
@@ -3,13 +3,13 @@ use std::ops::{Add, Sub};
 /// A point
 #[derive(Copy, Clone, Debug, Default)]
 pub struct Point {
-    pub x: isize,
-    pub y: isize,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Point {
     /// Create a new point
-    pub fn new(x: isize, y: isize) -> Self {
+    pub fn new(x: i32, y: i32) -> Self {
         Point { x: x, y: y }
     }
 }
